Sphinx docs.
- Added boilerplate for the various sections
- Changed theme to "classic"

ISSUES: Sphinx is breaking on vm_snapshots, vm, and host modules, for no apparent reason other than it dislikes me.
vm_snapshots is a good example of this, as it is the same as the other scripts, and the order it is inserted in scripts.rst doesn't help anything. *SIGH*.